pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "raccoon-cli"
dynamic = ["version"]
description = "Toolchain CLI for Botball robot development on Raspberry Pi (Wombat)"
readme = "README.md"
license = { text = "GPL-3.0-only" }
authors = [{ name = "Tobias Madlberger" }]
keywords = ["botball", "robotics", "raspberry-pi", "cli", "codegen"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Build Tools",
    "Topic :: Software Development :: Embedded Systems",
]
requires-python = ">=3.11"
dependencies = [
    "pyyaml",
    "ruamel.yaml>=0.17",
    "click>=8.0",
    "black",
    "rich>=10.0",
    "questionary>=2.0",
    "setuptools>=61.0",
    "wheel>=0.45.1",
    "jinja2>=3.0",
    "jinja2-time>=0.2",
    # Server dependencies (Pi-side)
    "fastapi>=0.100",
    "uvicorn[standard]>=0.20",
    "websockets>=10.0",
    # Client dependencies (Laptop-side)
    "paramiko>=4.0",
    "httpx>=0.24",
    "websocket-client>=1.0",
    "pyftpsync>=4.0",
    # IDE backend dependencies
    "libcst>=1.0.0",
    "pydantic>=2.0",
    "lcm",
    "raccoon-transport>=0.1",
    "raccoon-stubs",
]

[project.scripts]
raccoon = "raccoon_cli.cli:main"
raccoon-server = "raccoon_cli.server.cli:main"

[tool.setuptools.dynamic]
version = {attr = "raccoon_cli._version.__version__"}

[tool.setuptools.packages.find]
where = ["."]
include = ["raccoon_cli*"]

[tool.setuptools.package-data]
raccoon_cli = [
    "templates/**/*",
    "templates/**/*.jinja",
    "templates/**/*.py",
    "templates/**/*.sh",
    "templates/**/*.gitignore",
    "web-ide-dist/**/*",
]
